Move tests from Schema.test.ts to StateView.test.ts
