+++ test/Schema.test.ts
@@ -7,7 +7,7 @@
 // import "core-js";
 
 import * as assert from "assert";
-import { State, Player, DeepState, DeepMap, DeepChild, Position, assertDeepStrictEqualEncodeAll, createInstanceFromReflection, getEncoder, InheritanceRoot, createClientWithView, encodeMultiple } from "./Schema";
+import { State, Player, DeepState, DeepMap, DeepChild, Position, DeepEntity, assertDeepStrictEqualEncodeAll, createInstanceFromReflection, getEncoder } from "./Schema";
 import { Schema, ArraySchema, MapSchema, type, Metadata, $changes, Encoder, Decoder, SetSchema, schema, ToJSON, $refId } from "../src";
 import { getNormalizedType } from "../src/Metadata";
 
@@ -1015,87 +1015,6 @@ describe("Type: Schema", () => {
 
             decodedState.decode(encoded);
             assert.deepStrictEqual(decodedState.mapOfNumbers.toJSON(), { 'zero': 0, 'one': 1, 'two': 2, 'three': 3 });
-        });
-
-        it("should encode nested Schema", () => {
-            const state = new InheritanceRoot();
-            const encoder = getEncoder(state);
-
-            const client = createClientWithView(state);
-            client.view.add(state.parent);
-
-            // Initial encode: child property is undefined
-            encodeMultiple(encoder, state, [client]);
-
-            // Assign a child Schema instance.
-            state.parent.standardChild = new Position(1, 2, 3);
-
-            /**
-             * Encode an assignment of a child field:
-             * Its fields should be visible to the client, because it inherits visibility from its parent.
-             */
-            encodeMultiple(encoder, state, [client]);
-
-            assert.notStrictEqual(client.state.parent, undefined);
-            assert.notStrictEqual(client.state.parent.standardChild, undefined);
-            assert.strictEqual(client.state.parent.standardChild.x, 1);
-            assert.strictEqual(client.state.parent.standardChild.y, 2);
-            assert.strictEqual(client.state.parent.standardChild.z, 3);
-        });
-
-        it("should not encode nested Schema with @view", () => {
-            const state = new InheritanceRoot();
-            const encoder = getEncoder(state);
-
-            const client = createClientWithView(state);
-            client.view.add(state.parent);
-
-            // Initial encode: child property is undefined
-            encodeMultiple(encoder, state, [client]);
-
-            // Assign a child Schema that uses @view
-            state.parent.viewChild = new Position(1, 2, 3);
-
-            /**
-             * Encode an assignment of a child field:
-             * the child "viewChild" field (Position) is marked with @view,
-             * so it does not share visibility with its parent, and
-             * its fields are not encoded for the client.
-             */
-            encodeMultiple(encoder, state, [client]);
-
-            assert.notStrictEqual(client.state.parent, undefined);
-            assert.notStrictEqual(client.state.parent.viewChild, undefined);
-            assert.strictEqual(client.state.parent.viewChild.x, undefined);
-            assert.strictEqual(client.state.parent.viewChild.y, undefined);
-            assert.strictEqual(client.state.parent.viewChild.z, undefined);
-        });
-
-        it("should encode nested Schema wrapped in ArraySchema", () => {
-            const state = new InheritanceRoot();
-            const encoder = getEncoder(state);
-
-            const client = createClientWithView(state);
-            client.view.add(state.parent);
-
-            // Initial encode: child property is undefined
-            encodeMultiple(encoder, state, [client]);
-
-            // Assign a child Schema wrapped in an ArraySchema, to demonstrate this workaround.
-            state.parent.arrayChild.push(new Position(1, 2, 3));
-
-            /**
-             * Encode an assignment of a child field wrapped in an ArraySchema
-             * the child "arrayChild" field (Position) shares visibility with its parent,
-             * and its fields are encoded for the client.
-             */
-            encodeMultiple(encoder, state, [client]);
-
-            assert.notStrictEqual(client.state.parent, undefined);
-            assert.strictEqual(client.state.parent.arrayChild.length, 1);
-            assert.strictEqual(client.state.parent.arrayChild[0].x, 1);
-            assert.strictEqual(client.state.parent.arrayChild[0].y, 2);
-            assert.strictEqual(client.state.parent.arrayChild[0].z, 3);
         });
 
         describe("no changes", () => {
